pyproject.toml
[tool.poetry]
name = "graph-constrained-reasoning"
version = "0.1.0"
description = ""
authors = ["Linhao Luo <noreply@github.io>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.10"
transformers = "4.44"
accelerate = "^0.30.1"
peft = "^0.11.1"
deepspeed = "^0.14.2"
tiktoken = "^0.7.0"
openai = "^1.31.1"
httpx = ">=0.27,<0.28"
datasets = "^2.19.2"
python-dotenv = "^1.0.1"
marisa-trie = "^1.2.0"
scikit-learn = "^1.5.0"
bitsandbytes = "^0.43.1"
trl = "^0.9.4"
sentencepiece = "^0.2.0"
protobuf = "^5.27.1"
wandb = "^0.17.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
